Move protoloader classes to alongside other test helpers instead of tests
