+++ gapic-generator-java/src/test/java/com/google/api/generator/gapic/composer/grpc/GrpcServiceCallableFactoryClassComposerTest.java
@@ -18,6 +18,7 @@ import com.google.api.generator.gapic.model.GapicClass;
 import com.google.api.generator.gapic.model.GapicContext;
 import com.google.api.generator.gapic.model.Service;
 import com.google.api.generator.test.framework.Assert;
+import com.google.api.generator.test.protoloader.GrpcTestProtoLoader;
 import org.junit.Test;
 
 public class GrpcServiceCallableFactoryClassComposerTest {
