+++ gapic-generator-java/src/test/java/com/google/api/generator/gapic/composer/grpc/ServiceSettingsClassComposerTest.java
@@ -14,11 +14,11 @@
 
 package com.google.api.generator.gapic.composer.grpc;
 
-import com.google.api.generator.gapic.composer.common.TestProtoLoader;
 import com.google.api.generator.gapic.model.GapicClass;
 import com.google.api.generator.gapic.model.GapicContext;
 import com.google.api.generator.gapic.model.Service;
 import com.google.api.generator.test.framework.Assert;
+import com.google.api.generator.test.protoloader.TestProtoLoader;
 import java.util.Arrays;
 import java.util.Collection;
 import org.junit.Test;
