+++ gapic-generator-java/src/test/java/com/google/api/generator/test/protoloader/TestProtoLoader.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.google.api.generator.gapic.composer.common;
+package com.google.api.generator.test.protoloader;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
